Base class was added
