+++ js/components/PhonesCatalog.js
@@ -1,7 +1,8 @@
-export default class PhonesCatalog {
+import Component from '../Component.js';
+
+export default class PhonesCatalog extends Component {
   constructor(element, props) {
-    this.element = element;
-    this.props = props;
+    super(element, props);
 
     this.render();
 
@@ -15,8 +16,17 @@ export default class PhonesCatalog {
       const phoneId = link.dataset.phoneId;
       
       this.props.onPhoneSelected(phoneId)
-    })
+    });
+
+    this.on('click', 'PhoneLink', (event) => {
+      const phoneId = event.delegateTarget.dataset.phoneId;
+      this.props.onPhoneSelected(phoneId)
+    });
   }
+
+  
+  }
+
 
   render() {
     this.element.innerHTML = `
